refactor: move package managers scrapers to dedicated package
