+++ scrapers/src/main/java/nl/esciencecenter/rsd/scraper/package_manager/scrapers/NpmScraper.java
@@ -1,30 +1,31 @@
-// SPDX-FileCopyrightText: 2023 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
-// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2023 - 2024 Ewan Cahen (Netherlands eScience Center) <e.cahen@esciencecenter.nl>
+// SPDX-FileCopyrightText: 2023 - 2024 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package nl.esciencecenter.rsd.scraper.package_manager;
+package nl.esciencecenter.rsd.scraper.package_manager.scrapers;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
 
 import nl.esciencecenter.rsd.scraper.RsdResponseException;
+import nl.esciencecenter.rsd.scraper.Utils;
 
 import java.io.IOException;
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-public class PypiScraper implements PackageManagerScraper {
+public class NpmScraper implements PackageManagerScraper {
 
 	private final String packageName;
-	private static final Pattern urlPattern = Pattern.compile("https://pypi\\.org/project/([^/]+)/?");
+	private static final Pattern urlPattern = Pattern.compile("https://www\\.npmjs\\.com/package/(.+?)/?");
 
-	public PypiScraper(String url) {
+	public NpmScraper(String url) {
 		Objects.requireNonNull(url);
 		Matcher urlMatcher = urlPattern.matcher(url);
 		if (!urlMatcher.matches()) {
-			throw new RuntimeException("Invalid PyPi URL: " + url);
+			throw new RuntimeException("Invalid NPM URL: " + url);
 		}
 
 		packageName = urlMatcher.group(1);
@@ -37,7 +38,7 @@ public class PypiScraper implements PackageManagerScraper {
 
 	@Override
 	public Integer reverseDependencies() throws IOException, InterruptedException, RsdResponseException {
-		String data = PackageManagerScraper.doLibrariesIoRequest("https://libraries.io/api/pypi/" + packageName);
+		String data = PackageManagerScraper.doLibrariesIoRequest("https://libraries.io/api/npm/" + Utils.urlEncode(packageName));
 		JsonElement tree = JsonParser.parseString(data);
 		return tree.getAsJsonObject().getAsJsonPrimitive("dependents_count").getAsInt();
 	}
